Add Cart#isEmpty property, <stripe-checkout> uses `cart` to set disabled
This change works around an issue that we discovered with tracking args.
See: https://github.com/glimmerjs/glimmer-component/issues/66

The issue was: The "Pay by credit card" button in stripe-checkout was
initially out of sync if the user clicks a sticker's "Add to cart"
button for the very first time.